License Plate Recognition/requirements.txt
opencv-python-headless
torch
torchvision
Pillow
ultralytics
transformers
